센터 등록 아이디 중복검사 비동기 통신 및 형식 검사 구현
