+++ semi/src/main/java/semi/servlet/ajax/CenterTrainerListServlet.java
@@ -11,6 +11,7 @@ import javax.servlet.http.HttpServletResponse;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
+import semi.servlet.DtoDao.CenterDao;
 import semi.servlet.DtoDao.TrainerDao;
 import semi.servlet.DtoDao.TrainerDto;
 @WebServlet(urlPatterns = "/ajax/center-trainer.kh")
